updated speed_profile.py
First correction by Codex to track definition. Enforced that a point marked as straight will be straight to the next point. It's noted that the sample track provided does not allow the corner radius to be specified. If an intermediate corner point is added, the first part of the corner has an incorrect radius. sample_track.csv updated to allow Codex to investigate.